Fix some bugs
Small change for API, user can see the result set but not able to change. The staff would be able to see and change the survey. (if we don't allow the user see, we need to take down the API for survey, and only allow super user to see and change, or we need to set a redirect after login page for staff). Also, fixed some bug in users.